draft classification script
