+++ .github/workflows/python-release.yml
@@ -1,0 +1,44 @@
+name: Python package release
+
+on:
+  workflow_dispatch:
+    inputs:
+      version:
+        description: "Version to release (must match pyproject.toml)"
+        required: false
+
+jobs:
+  test:
+    uses: ./.github/workflows/python-test.yml
+
+  publish:
+    needs: test
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/project/licenses-aas/
+    permissions:
+      id-token: write
+      contents: read
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install build backend
+        run: |
+          python -m pip install --upgrade pip
+          pip install build
+
+      - name: Build distributions
+        run: |
+          python -m build
+
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          verbose: true
