+++ .github/workflows/python-test.yml
@@ -1,0 +1,30 @@
+name: Python package tests
+
+on:
+  push:
+    branches: [ main, master ]
+  pull_request:
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install package
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+
+      - name: Run basic import test
+        run: |
+          python -c "import licensing, licensing.classify.classify_license, licensing.inspect_licenses, licensing.merge_spdx_with_choosealicense; print('ok')"
