+++ pyproject.toml
@@ -1,0 +1,44 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "licenses-aas"
+version = "0.1.0"
+description = "Developer-friendly tools and data for exploring and classifying software and data licenses."
+readme = "README.md"
+license = { text = "Apache License, Version 2.0" }
+requires-python = ">=3.9"
+dependencies = [
+  "requests",
+  "pandas",
+  "pyyaml",
+  "openai",
+]
+authors = [
+  { name = "MobilityData" },
+]
+maintainers = [
+  { name = "MobilityData" },
+]
+classifiers = [
+  "Intended Audience :: Developers",
+  "License :: OSI Approved :: Apache Software License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Topic :: Software Development :: Libraries :: Python Modules",
+]
+
+[project.urls]
+Homepage = "https://github.com/MobilityData/licenses-aas"
+Repository = "https://github.com/MobilityData/licenses-aas"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+packages = ["licensing", "licensing.classify"]
