+++ src/licensing/__init__.py
@@ -1,0 +1,5 @@
+"""Core tools for working with software and data licenses.
+
+Subpackages:
+- classify: helpers for LLM-based classification and tag heuristics
+"""
