+++ src/licensing/classify/__init__.py
@@ -1,0 +1,6 @@
+"""Classification helpers for licenses.
+
+Modules:
+- classify_license: LLM-based classification workflows
+- license_tags: tag heuristics and registry utilities
+"""
